\name{GWASinlps-package}

\alias{GWASinlps-package}

\docType{package}

\title{Nonlocal prior based iterative SNP selection tool for genome-wide association study data}

\description{Performs variable selection with data from Genome-wide association studies (GWAS) combining, in an iterative variable selection framework, the computational efficiency of the screen-and-select approach based on some association learning and the parsimonious uncertainty quantification provided by the use of nonlocal priors, as described in Sanyal et al. (2018).}

\details{
\tabular{ll}{
Package: \tab GWASinlps\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2017-07-06\cr
License: \tab GPL (>= 2)\cr
}
%~~ An overview of how to use the package, including the most ~~
%~~ important functions ~~
}

\author{Nilotpal Sanyal <nilotpal.sanyal@gmail.com>

Maintainer: Nilotpal Sanyal <nilotpal.sanyal@gmail.com>
}


\references{
Sanyal et al. (2018), "GWASinlps: Nonlocal prior based iterative SNP selection tool for genome-wide association studies".
}
