\name{GWASinlps-package}

\alias{GWASinlps-package}

\docType{package}

\title{Nonlocal prior based SNP selection tool for genome-wide association study data}

\description{Performs variable selection regression and prediction with data from Genome-wide association study (GWAS) using non-local prior based Bayesian generalized linear model as described in Sanyal et al. (2017)(submitted).}

\details{
\tabular{ll}{
Package: \tab GWASinlps\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-07-06\cr
License: \tab GPL (>= 2)\cr
}
%~~ An overview of how to use the package, including the most ~~
%~~ important functions ~~
}

\author{Nilotpal Sanyal <nisanyal@ucsd.edu>

Maintainer: Nilotpal Sanyal <nisanyal@ucsd.edu>
}


\references{
Sanyal et al. (2017), "GWASinlps: Nonlocal prior based SNP selection tool for genome-wide association studies" (submitted to Bioinformatics).
}
